backend/requirements.txt
django
djangorestframework
django-cors-headers
Pillow
mongoengine
pymongo
django-rest-framework-mongoengine
